chore: .csv files are default for exporting
Waiting a little bit before merging this into main. When trying to
open this in Quick Look, it seems broken. Though when I open it
in LibreOffice Calc, it opens correctly. I'm going to try to take
more time to fix the Quick Look version a bit more before merging.